Basic evaluation function & "Negamax"
From-psuedocode interpretation of Negamax. Issue is that we're
generating and cloning new board states. As opposed to real Negamax
where we take a state, make a move, negamax it, then undo the move
before making a new move.

Also added basic Evaluation function to Board class. Evaluation can be
tweaked as needed.