Rough NegaMax
Beginning of NegaMax implementation. Redid all of the move generation. A
Move is now a saved state of what a move would do in the game. Instead
of generating board states, we generate moves. All Boards have a running
stack of moves done on them. The moves are actually applied to the board
with "MakeMove" and can be undone with "UndoMove", which manipulate the
stack (push and pop respectively). TODOs are noted in Board,
particularly with Phil's advanced MovePiece from the older function. Now
MovePieces simply return Move objects, which probably can be refactored
out maybe for speed and less redundant code.